pyproject.toml
[project]
name = "dental"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.1.3",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "protobuf>=4.21.0,<6.0.0",
    "scikit-learn>=1.7.0",
    "setuptools<81",
    "six>=1.17.0",
    "tensorflow>=2.19.0",
    "tensorflow-hub>=0.16.1",
]
